pyproject.toml
[project]
name = "ai-research-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "langchain>=1.2.13",
    "langchain-mcp-adapters>=0.2.2",
    "langchain-ollama>=1.0.1",
    "langgraph>=1.1.3",
    "python-dotenv>=1.2.2",
]
